feat: migrate list.integration.test.ts to v2 mock system
Replaces legacy DI test pattern with v2 TestContainerBuilder and predefined scenarios. Adds comprehensive test coverage including:

- All original test cases migrated to v2 patterns
- Calendar argument handling tests
- Pretty-json format tests
- Enhanced error handling scenarios
- Consistent use of testScenarios helpers
